requirements.txt
# Google Gen AI SDK for Veo API (expensive - ~$1.75/video)
google-genai>=0.1.0
google-cloud-aiplatform>=1.38.0,<2.0.0
google-cloud-storage>=2.10.0,<3.0.0
google-auth>=2.23.0,<3.0.0

# Replicate API - cheapest alternative (~$0.05/video)
replicate>=0.25.0

# OpenAI Sora API - medium cost (~$0.50-2.50/video)
openai>=1.51.0

# HTTP requests
requests>=2.31.0,<3.0.0
urllib3>=1.26.0,<2.0.0  # Compatible with Python 3.9
aiohttp>=3.8.0,<4.0.0

# Video processing
ffmpeg-python>=0.2.0

# TTS - ElevenLabs API
elevenlabs>=0.2.27,<1.0.0

# Environment variables
python-dotenv>=1.0.0

# CLI interface
click>=8.1.7
rich>=13.7.0

# JSON handling
pydantic>=2.5.0,<3.0.0

# Claude API for video analysis
anthropic>=0.40.0

# Fix for importlib.metadata compatibility with Python 3.9
importlib-metadata>=4.0.0; python_version < "3.10"
